+++ dev/HR_Mudol/DTO/ConstraintDTO.java
@@ -4,19 +4,19 @@ package HR_Mudol.DTO;
  * Data Transfer Object representing a Constraint for an employee's availability.
  */
 public class ConstraintDTO {
-    private int empID;
+    private long empID;
     private String explanation;
     private String day;   // e.g., "SUNDAY"
     private String type;  // e.g., "MORNING"
 
-    public ConstraintDTO(int ID, String explanation, String day, String type) {
+    public ConstraintDTO(long ID, String explanation, String day, String type) {
         this.empID=ID;
         this.explanation = explanation;
         this.day = day;
         this.type = type;
     }
 
-    public int getEmpID() {
+    public long getEmpID() {
         return empID;
     }
