+++ dev/HR_Mudol/DTO/EmployeeDTO.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class EmployeeDTO {
-    private int employeeId;
+    private long employeeId;
     private String fullName;
     private String password;
     private String bankAccount;
@@ -26,7 +26,7 @@ public class EmployeeDTO {
     private List<ConstraintDTO> eveningConstraints;
     private List<ConstraintDTO> lockedConstraints;
 
-    public EmployeeDTO(int employeeId, String fullName, String password, String bankAccount, int salary,
+    public EmployeeDTO(long employeeId, String fullName, String password, String bankAccount, int salary,
                        LocalDate startDate, int minDayShift, int minEveningShift, int sickDays, int daysOff) {
         this.employeeId = employeeId;
         this.fullName = fullName;
@@ -45,7 +45,7 @@ public class EmployeeDTO {
         this.lockedConstraints=List.of();;
     }
 
-    public EmployeeDTO(int employeeId, String fullName, String password, String bankAccount, int salary,
+    public EmployeeDTO(long employeeId, String fullName, String password, String bankAccount, int salary,
                        LocalDate startDate, int minDayShift, int minEveningShift, int sickDays, int daysOff,
                        List<Integer> relevantRoleIds, List<ConstraintDTO> weeklyConstraints,List<ConstraintDTO> eveningConstraints,List<ConstraintDTO> lockedConstraints , List<ConstraintDTO> morningConstraints) {
         this.employeeId = employeeId;
@@ -66,7 +66,7 @@ public class EmployeeDTO {
 
     }
 
-    public int getEmployeeId() { return employeeId; }
+    public long getEmployeeId() { return employeeId; }
     public String getFullName() { return fullName; }
     public String getPassword() { return password; }
     public String getBankAccount() { return bankAccount; }
